requirements.txt
mysql-connector-python==8.0.23
pip==21.0.1
protobuf==3.15.7
setuptools==54.2.0
six==1.15.0
Flask~=1.1.2